// src/app/models/daily-weather.model.ts
export class DailyWeather {
    date: string;
    precipitation_sum: number;
    precipitation_probability_max: number;
    weather_code: number;
    temperature_2m_max: number;
    temperature_2m_min: number;
    active: boolean;
    time: string;
    sunrise: string;
    sunset: string;
    wind_direction_10m: number;
    wind_gusts_10m_min: number;
    wind_gusts_10m_max: number;
}